separate index mapping class and box class
